requirements.txt
argcomplete
clint
colorama
nutcli>=1.1
pyyaml
requests
requests-toolbelt
